+++ tests/integration/components/modal-dialog-test.js
@@ -10,6 +10,7 @@ import {
   render,
   settled,
   focus,
+  triggerEvent,
   triggerKeyEvent,
   click
 } from '@ember/test-helpers';
@@ -110,9 +111,7 @@ module('modal-dialog', function (hooks) {
         return deferred.promise;
       };
 
-      this.inserted = () => {
-        assert.step('inserted');
-      };
+      this.inserted = () => assert.step('inserted');
 
       await render(hbs`
         <ModalDialog
@@ -206,7 +205,7 @@ module('modal-dialog', function (hooks) {
     test('yielded close', async function (assert) {
       assert.expect(3);
 
-      this.close = () => assert.step('close');
+      this.close = () => assert.step('closed');
 
       await render(hbs`
         <ModalDialog @onClose={{this.close}} as |modal|>
@@ -223,7 +222,7 @@ module('modal-dialog', function (hooks) {
       await waitForAnimation('.modal-dialog');
 
       assert.verifySteps(
-        ['close'],
+        ['closed'],
         'close action is fired after the hide animation'
       );
     });
@@ -255,7 +254,7 @@ module('modal-dialog', function (hooks) {
       await waitForAnimation('.modal-dialog');
 
       assert.verifySteps(
-        ['close'],
+        ['closed'],
         'close action is fired after the hide animation'
       );
     });
@@ -270,7 +269,6 @@ module('modal-dialog', function (hooks) {
       `);
 
       await click('button');
-
       await waitForAnimation('.modal-dialog');
 
       assert.ok(true, 'does not blow up if onClose is not a function');
@@ -293,38 +291,38 @@ module('modal-dialog', function (hooks) {
     });
   });
 
-  module('escaping', function () {
+  module('escaping', function (hooks) {
+    hooks.beforeEach(function (assert) {
+      this.close = () => assert.step('closed');
+    });
+
     // Modal dialogs that do not close when escape is pressed add a class name
     // to the modal, so you can add a suitable animation to inform the user
     // that their action was denied. This is useful for preventing accidental
     // data loss, if a user has entered text into a modal, then hits escape
     // without pressing Save for example.
 
-    test('allowed', async function (assert) {
+    test('pressing escape (allowed)', async function (assert) {
       assert.expect(2);
-
-      this.close = () => assert.step('close');
 
       await render(hbs`
         <ModalDialog
           @escapable={{true}}
-          @onClose={{this.close}} />
+          @onClose={{this.close}}
+        />
       `);
 
       await triggerKeyEvent('.modal-dialog', 'keydown', 27); // Escape
-
       await waitForAnimation('.modal-dialog');
 
       assert.verifySteps(
-        ['close'],
+        ['closed'],
         'escapable modal dialogs will close when escape is pressed'
       );
     });
 
-    test('not allowed', async function (assert) {
+    test('pressing escape (not allowed)', async function (assert) {
       assert.expect(3);
-
-      this.close = () => assert.step('close');
 
       await render(hbs`
         <ModalDialog @onClose={{this.close}} />
@@ -349,6 +347,47 @@ module('modal-dialog', function (hooks) {
         );
 
       assert.verifySteps([], 'close action is not fired');
+    });
+
+    test('clicking outside to escape (not allowed)', async function (assert) {
+      assert.expect(1);
+
+      await render(hbs`<ModalDialog @onClose={{this.close}} />`);
+
+      await click('.modal-dialog');
+      await waitForAnimation('.modal-dialog');
+
+      assert.verifySteps([], 'is not escapable');
+    });
+
+    test('clicking outside to escape (allowed)', async function (assert) {
+      assert.expect(2);
+
+      await render(
+        hbs`<ModalDialog @escapable={{true}} @onClose={{this.close}} />`
+      );
+
+      await click('.modal-dialog');
+      await waitForAnimation('.modal-dialog');
+
+      assert.verifySteps(
+        ['closed'],
+        'clicking outside the modal dialog box closes the modal'
+      );
+    });
+
+    test('clicking inside and releasing outside', async function (assert) {
+      assert.expect(1);
+
+      await render(
+        hbs`<ModalDialog @escapable={{true}} @onClose={{this.close}} />`
+      );
+
+      await triggerEvent('.modal-dialog__box', 'mousedown');
+      await triggerEvent('.modal-dialog', 'mouseup');
+      await waitForAnimation('.modal-dialog');
+
+      assert.verifySteps([], 'does not close');
     });
   });
 
@@ -508,7 +547,6 @@ module('modal-dialog', function (hooks) {
         .isFocused('modal is focused to respond the keyboard');
 
       await click('.internal');
-
       await waitForAnimation('.modal-dialog');
 
       assert